Refactor tests
phovea/generator-phovea#425
