+++ test/add-extension.test.js
@@ -4,9 +4,9 @@ const path = require('path');
 const assert = require('yeoman-assert');
 const helpers = require('yeoman-test');
 const fse = require('fs-extra');
-const TestUtils = require('./TestUtils');
+const TestUtils = require('./test-utils/TestUtils');
 const {template} = require('lodash');
-const dependencies = require('./generator-dependencies');
+const dependencies = require('./test-utils/generator-dependencies');
 
 describe('add a web extension to a web library', () => {
     const tdpViewTmpl = template(fse.readFileSync(TestUtils.templatePath('add-extension', 'tdpView.tmpl.ts')))({moduleName: 'CustomView'});
