+++ test/init-lib.test.js
@@ -5,10 +5,10 @@ const assert = require('yeoman-assert');
 const helpers = require('yeoman-test');
 const rimraf = require('rimraf');
 const fse = require('fs-extra');
-const TestUtils = require('./TestUtils');
+const TestUtils = require('./test-utils/TestUtils');
 const {template} = require('lodash');
 const SpawnUtils = require('../utils/SpawnUtils');
-const dependencies = require('./generator-dependencies');
+const dependencies = require('./test-utils/generator-dependencies');
 
 /**
  * Directory name to run the generator
